+++ R/generics.R
@@ -619,11 +619,11 @@ tensor_ops <- function(a, b) UseMethod("tensor_ops")
 }
 
 
-#' @export
-"round.torch.Tensor" <- function(input, ...) {
-    # round: Returns a new tensor with each of the elements of input rounded to the closest integer.
-    torch$round(input, ...)
-}
+#' #' @export
+#' "round.torch.Tensor" <- function(input) {
+#'     # round: Returns a new tensor with each of the elements of input rounded to the closest integer.
+#'     torch$round(input)
+#' }
 
 
 #' Dot product of two tensors